backend/app/utils/cli.py
from typing import Optional, Callable, Dict
import os
import time
from dotenv import load_dotenv

load_dotenv()

PROJECT_BASE_DIR = os.getenv("PROJECT_BASE_DIR", "./projects")

async def createAPI(
    description: str,
    output_dir: str,
    broadcast_callback: Optional[Callable] = None,
    use_docker: bool = True,
    use_nginx: bool = False
) -> Dict:
    try:
        # TODO: Implement actual CLI call to oneShotCodeGen
        # This is a placeholder for the actual implementation
        if broadcast_callback:
            await broadcast_callback("Starting app generation...")
            await broadcast_callback("Generating code...")
            await broadcast_callback("Building Docker container...")

        return {
            "status": "success",
            "message": "App created successfully",
            "output_dir": output_dir,
            "preview_url": f"http://localhost:3000/{os.path.basename(output_dir)}",
            "use_cases": {}  # TODO: Add actual use cases
        }
    except Exception as e:
        return {
            "status": "error",
            "message": str(e)
        }

async def editAPI(
    project_dir: str,
    description: str,
    broadcast_callback: Optional[Callable] = None,
    use_docker: bool = True,
    use_nginx: bool = False
) -> Dict:
    try:
        # TODO: Implement actual CLI call to oneShotCodeGen
        backup_dir = f"{project_dir}_backup_{int(time.time())}"
        
        if broadcast_callback:
            await broadcast_callback("Starting app modification...")
            await broadcast_callback("Applying changes...")
            await broadcast_callback("Rebuilding Docker container...")

        return {
            "status": "success",
            "message": "App modified successfully",
            "backup_dir": backup_dir,
            "preview_url": f"http://localhost:3000/{os.path.basename(project_dir)}",
            "use_cases": {}  # TODO: Add actual use cases
        }
    except Exception as e:
        return {
            "status": "error",
            "message": str(e)
        }

async def revertAPI(
    project_dir: str,
    backup_dir: str,
    broadcast_callback: Optional[Callable] = None,
    use_docker: bool = True,
    use_nginx: bool = False
) -> Dict:
    try:
        # TODO: Implement actual CLI call to oneShotCodeGen
        if broadcast_callback:
            await broadcast_callback("Starting reversion process...")
            await broadcast_callback("Restoring from backup...")
            await broadcast_callback("Rebuilding Docker container...")

        return {
            "status": "success",
            "message": "App reverted successfully",
            "preview_url": f"http://localhost:3000/{os.path.basename(project_dir)}"
        }
    except Exception as e:
        return {
            "status": "error",
            "message": str(e)
        } 