backend/requirements.txt
fastapi==0.115.7
uvicorn==0.34.0
python-dotenv==1.0.1
websockets==13.0.0
python-multipart==0.0.20
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
aiofiles==24.1.0
click==8.1.8
python-jwt==4.1.0
cryptography==44.0.0
httpx
typer[all]==0.9.0
numpy==2.0.2
pydantic==2.10.3
outlines==0.1.11
rich==13.7.0
jinja2==3.1.3
supabase==2.12.0
asyncio==3.4.3
aiohttp==3.11.1
rich-click==1.7.3
inquirer==3.2.4
colorama==0.4.6
openai==1.57.1
setuptools>=65.0.0
wheel>=0.36.0
psycopg2-binary==2.9.10
pyjwt==2.10.0
nest-asyncio==1.6.0
